+++ collectors/nginx_test.go
@@ -37,7 +37,7 @@ func TestGetNginxStatus(t *testing.T) {
 	for _, test := range tests {
 		g.Describe("getNginxStatus()", func() {
 			g.It(test.TestDescription, func() {
-				result := getNginxStatus(fakeConfig, make(chan map[string]interface{}, 1), test.HTTPRunner)
+				result := getNginxStatus(fakeConfig, make(chan []map[string]interface{}, 1), test.HTTPRunner)
 				g.Assert(reflect.DeepEqual(result, string(test.HTTPRunner.Data))).Equal(true)
 			})
 		})
@@ -47,22 +47,25 @@ func TestGetNginxStatus(t *testing.T) {
 func TestScrapeStatus(t *testing.T) {
 	g := goblin.Goblin(t)
 
+	resultSlice := make([]map[string]interface{}, 1)
+	resultSlice[0] = map[string]interface{}{
+		"nginx.net.connections": "2",
+		"nginx.net.accepts":     "29",
+		"nginx.net.handled":     "29",
+		"nginx.net.requests":    "31",
+		"nginx.net.writing":     "1",
+		"nginx.net.waiting":     "1",
+		"nginx.net.reading":     "0",
+	}
+
 	var tests = []struct {
 		Data            string
-		ExpectedResult  map[string]interface{}
+		ExpectedResult  []map[string]interface{}
 		TestDescription string
 	}{
 		{
-			Data: "Active connections: 2 \nserver accepts handled requests\n 29 29 31 \nReading: 0 Writing: 1 Waiting: 1 ",
-			ExpectedResult: map[string]interface{}{
-				"nginx.net.connections": "2",
-				"nginx.net.accepts":     "29",
-				"nginx.net.handled":     "29",
-				"nginx.net.requests":    "31",
-				"nginx.net.writing":     "1",
-				"nginx.net.waiting":     "1",
-				"nginx.net.reading":     "0",
-			},
+			Data:            "Active connections: 2 \nserver accepts handled requests\n 29 29 31 \nReading: 0 Writing: 1 Waiting: 1 ",
+			ExpectedResult:  resultSlice,
 			TestDescription: "Successfully scrape given status page",
 		},
 	}