+++ collectors/types.go
@@ -4,9 +4,9 @@ var CollectorArray map[string]Collector
 
 // As we develop new collectors, add to this list here, so that our dispatcher
 // knows about them
-//		"nginx":    NginxCollector,
 func init() {
 	CollectorArray = map[string]Collector{
+		"nginx":    NginxCollector,
 		"rabbitmq": RabbitmqCollector,
 	}
 
@@ -26,22 +26,29 @@ func init() {
 type Collector func(config Config, stats chan<- []map[string]interface{})
 
 type Config struct {
-	AppName        string
-	NginxConfig    NginxConfig
-	RabbitmqConfig RabbitmqConfig
+	AppName        string                  `yaml:"appname"`
+	NewRelicKey    string                  `yaml:"newrelickey"`
+	DefaultDelayMS int                     `yaml:"defaultdelayms"`
+	Collectors     map[string]CommonConfig `yaml:"collectors"`
 }
 
+type CommonConfig struct {
+	Enabled         bool            `yaml:"enabled" required:"true"`
+	DelayMS         int             `yaml:"delayms"`
+	CollectorConfig CollectorConfig `yaml:"collectorconfig"`
+}
+
+type CollectorConfig interface{}
+
 type NginxConfig struct {
-	Enabled         bool
-	NginxListenPort string
-	NginxStatusURI  string
-	NginxStatusPage string
+	NginxListenPort string `yaml:"nginxlistenport"`
+	NginxStatusURI  string `yaml:"nginxstatusuri"`
+	NginxStatusPage string `yaml:"nginxstatuspage"`
 }
 
 type RabbitmqConfig struct {
-	Enabled  bool
-	User     string
-	Password string
-	Port     string
-	Host     string
+	RabbitMQUser     string `yaml:"rabbitmquser"`
+	RabbitMQPassword string `yaml:"rabbitmqpassword"`
+	RabbitMQPort     string `yaml:"rabbitmqport"`
+	RabbitMQHost     string `yaml:"rabbitmqhost"`
 }