+++ CHANGELOG.md
@@ -3,6 +3,14 @@ go-api-skeleton CHANGELOG
 
 This file is used to list changes made in each version of go-newrelic-plugin.
 
+# 0.0.3
+
+Michael Dunton - adding support and Multiple Stats from a plugin and adding Rabbitmq
+
+# 0.0.2
+
+Tom Barber - adding support and structs for reading config file
+
 # 0.0.1
 
 Michael Dunton - initial change log commit of go-newrelic-plugin